requirements.txt
# WTI Oil Trading Bot - Python Dependencies

# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# MetaTrader integration
MetaTrader5>=5.0.45

# Timezone handling
pytz>=2023.3

# Logging and utilities
python-dateutil>=2.8.2

# Optional: For backtesting and visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Development dependencies
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
